try to fix without correction

+++ common/hardwareAbstraction/hardwareAbstraction.c
@@ -151,7 +151,7 @@ int GetSensorValues(uint16_t *p_ptISensors)
 ////////////////////////////////////////////////////////////
 /// @brief Hardware Abstraction Layer get motor encoder values
 ////////////////////////////////////////////////////////////
-int GetMotorEncoderValues(uint16_t *p_ptIMotors)
+int GetMotorEncoderValues(int32_t *p_ptIMotors)
 {
     X_ASSERT(t_bHardwareAbstractionInitialized == true);
 
@@ -161,7 +161,7 @@ int GetMotorEncoderValues(uint16_t *p_ptIMotors)
         int l_iValue = mrpiz_motor_encoder_get(t_HardwareAbstraction.t_iMotors[i]);
         if (l_iValue != -1)
         {
-            p_ptIMotors[i] = (int16_t)l_iValue;
+            p_ptIMotors[i] = (int32_t)l_iValue;
         }
 
         else
